style: fix cargo fmt and clippy issues from CI
Run cargo fmt and resolve all clippy warnings including:
- while_let_loop, collapsible_if, derivable_impls
- should_implement_trait, len_without_is_empty
- print_literal, needless_return, needless_borrow
- field_reassign_with_default, too_many_arguments
- dead_code, unnecessary_cast, if_same_then_else

+++ src/tui/components/help.rs
@@ -206,7 +206,7 @@ pub fn render_input(frame: &mut Frame, title: &str, input: &str) {
 /// Render a selection dialog overlay.
 pub fn render_select(frame: &mut Frame, title: &str, options: &[String], selected: usize) {
     let height = (options.len() as u16 + 5).min(20);
-    let area = centered_rect(50, height.max(8) as u16, frame.area());
+    let area = centered_rect(50, height.max(8), frame.area());
     frame.render_widget(Clear, area);
 
     let mut lines = vec![Line::raw("")];
